--- src/App.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router-dom';
-
-function App() {
-  return (
-    <div>
-      TEST
-    </div>
-  );
-}
-
-export default App;
